+++ source/components/radio-buttons/index.js
@@ -1,0 +1,3 @@
+// n8n radio buttons
+import "./radio-buttons.less"
+import "./radio-buttons.js"