+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 5.1.7
 
++ [x] Added component `radio-buttons` (n8n style)
++ [x] Fixed component `splitter` for save state 
 + [x] Updated docs for a lot of components